fix #122037
