+++ src/vs/workbench/contrib/terminal/browser/terminalActions.ts
@@ -42,20 +42,6 @@ import { IRemoteAgentService } from 'vs/workbench/services/remote/common/remoteA
 export const switchTerminalActionViewItemSeparator = '─────────';
 export const switchTerminalShowTabsTitle = localize('showTerminalTabs', "Show Tabs");
 
-const enum ContextMenuGroup {
-	Create = '1_create',
-	Edit = '2_edit',
-	Clear = '3_clear',
-	Kill = '4_kill',
-	Config = '5_config'
-}
-
-export const enum ContextMenuTabsGroup {
-	Default = '1_create_default',
-	Profile = '2_create_profile',
-	Configure = '3_configure'
-}
-
 async function getCwdForSplit(configHelper: ITerminalConfigHelper, instance: ITerminalInstance, folders?: IWorkspaceFolder[], commandService?: ICommandService): Promise<string | URI | undefined> {
 	switch (configHelper.config.splitCwd) {
 		case 'workspaceRoot':
@@ -204,13 +190,7 @@ export function registerTerminalActions() {
 			await terminalService.showPanel(true);
 		}
 	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalTabsWidgetEmptyContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.NEW_WITH_PROFILE,
-			title: localize('workbench.action.terminal.newWithProfile.short', "New Terminal With Profile")
-		},
-		group: ContextMenuGroup.Create
-	});
+
 
 	registerAction2(class extends Action2 {
 		constructor() {
@@ -226,14 +206,6 @@ export function registerTerminalActions() {
 			const terminalService = accessor.get(ITerminalService);
 			terminalService.showTabs();
 		}
-	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalContainerContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.SHOW_TABS,
-			title: localize('workbench.action.terminal.showsTabs', "Show Tabs")
-		},
-		when: ContextKeyExpr.not('config.terminal.integrated.tabs.enabled'),
-		group: ContextMenuGroup.Config
 	});
 
 	registerAction2(class extends Action2 {
@@ -753,11 +725,7 @@ export function registerTerminalActions() {
 				title: { value: localize('workbench.action.terminal.changeIcon', "Change Icon..."), original: 'Change Icon...' },
 				f1: true,
 				category,
-				precondition: KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED,
-				menu: {
-					id: MenuId.TerminalSingleTabContext,
-					group: ContextMenuGroup.Edit
-				}
+				precondition: KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED
 			});
 		}
 		async run(accessor: ServicesAccessor) {
@@ -771,11 +739,7 @@ export function registerTerminalActions() {
 				title: { value: localize('workbench.action.terminal.changeIcon', "Change Icon..."), original: 'Change Icon...' },
 				f1: false,
 				category,
-				precondition: ContextKeyExpr.and(KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED, KEYBINDING_CONTEXT_TERMINAL_TABS_SINGULAR_SELECTION),
-				menu: {
-					id: MenuId.TerminalTabsWidgetContext,
-					group: ContextMenuGroup.Edit
-				}
+				precondition: ContextKeyExpr.and(KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED, KEYBINDING_CONTEXT_TERMINAL_TABS_SINGULAR_SELECTION)
 			});
 		}
 		async run(accessor: ServicesAccessor) {
@@ -789,11 +753,7 @@ export function registerTerminalActions() {
 				title: { value: localize('workbench.action.terminal.rename', "Rename..."), original: 'Rename...' },
 				f1: true,
 				category,
-				precondition: KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED,
-				menu: {
-					id: MenuId.TerminalSingleTabContext,
-					group: ContextMenuGroup.Edit
-				}
+				precondition: KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED
 			});
 		}
 		async run(accessor: ServicesAccessor) {
@@ -816,16 +776,6 @@ export function registerTerminalActions() {
 					weight: KeybindingWeight.WorkbenchContrib
 				},
 				precondition: ContextKeyExpr.and(KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED, KEYBINDING_CONTEXT_TERMINAL_TABS_SINGULAR_SELECTION),
-				menu: [
-					{
-						id: MenuId.TerminalTabsWidgetContext,
-						group: ContextMenuGroup.Edit
-					},
-					{
-						id: MenuId.TerminalContainerContext,
-						group: ContextMenuGroup.Edit
-					}
-				]
 			});
 		}
 		async run(accessor: ServicesAccessor) {
@@ -1350,23 +1300,6 @@ export function registerTerminalActions() {
 					when: KEYBINDING_CONTEXT_TERMINAL_FOCUS
 				},
 				icon: Codicon.splitHorizontal,
-				menu: [
-					{
-						id: MenuId.ViewTitle,
-						group: 'navigation',
-						order: 2,
-						when: ContextKeyAndExpr.create([
-							ContextKeyEqualsExpr.create('view', TERMINAL_VIEW_ID),
-							ContextKeyExpr.not('config.terminal.integrated.tabs.enabled')
-						]),
-					}, {
-						id: MenuId.TerminalSingleTabContext,
-						group: ContextMenuGroup.Create
-					}, {
-						id: MenuId.TerminalContainerContext,
-						group: ContextMenuGroup.Create
-					}
-				],
 				description: {
 					description: 'workbench.action.terminal.split',
 					args: [{
@@ -1398,10 +1331,6 @@ export function registerTerminalActions() {
 				f1: false,
 				category,
 				precondition: KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED,
-				menu: {
-					id: MenuId.TerminalTabsWidgetContext,
-					group: ContextMenuGroup.Create
-				},
 				keybinding: {
 					primary: KeyMod.CtrlCmd | KeyMod.Shift | KeyCode.KEY_5,
 					mac: {
@@ -1422,13 +1351,7 @@ export function registerTerminalActions() {
 			}
 		}
 	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalTabInlineActions, {
-		command: {
-			id: TERMINAL_COMMAND_ID.SPLIT_INSTANCE,
-			title: localize('workbench.action.terminal.splitInstance', "Split Terminal"),
-		},
-		group: ContextMenuGroup.Create
-	});
+
 	registerAction2(class extends Action2 {
 		constructor() {
 			super({
@@ -1466,12 +1389,7 @@ export function registerTerminalActions() {
 					mac: { primary: KeyMod.CtrlCmd | KeyCode.KEY_A },
 					weight: KeybindingWeight.WorkbenchContrib,
 					when: KEYBINDING_CONTEXT_TERMINAL_FOCUS
-				}],
-				menu: {
-					id: MenuId.TerminalContainerContext,
-					group: ContextMenuGroup.Edit,
-					order: 3
-				}
+				}]
 			});
 		}
 		run(accessor: ServicesAccessor) {
@@ -1530,20 +1448,6 @@ export function registerTerminalActions() {
 			await terminalService.showPanel(true);
 		}
 	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalTabsWidgetEmptyContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.NEW,
-			title: localize('workbench.action.terminal.new.short', "New Terminal")
-		},
-		group: ContextMenuGroup.Create
-	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalContainerContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.NEW,
-			title: localize('workbench.action.terminal.new.short', "New Terminal")
-		},
-		group: ContextMenuGroup.Create
-	});
 	registerAction2(class extends Action2 {
 		constructor() {
 			super({
@@ -1574,20 +1478,7 @@ export function registerTerminalActions() {
 			});
 		}
 	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalContainerContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.KILL,
-			title: localize('workbench.action.terminal.kill.short', "Kill Terminal")
-		},
-		group: ContextMenuGroup.Kill
-	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalSingleTabContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.KILL,
-			title: localize('workbench.action.terminal.kill.short', "Kill Terminal")
-		},
-		group: ContextMenuGroup.Kill
-	});
+
 	registerAction2(class extends Action2 {
 		constructor() {
 			super({
@@ -1598,10 +1489,6 @@ export function registerTerminalActions() {
 				f1: false,
 				category,
 				precondition: ContextKeyExpr.or(KEYBINDING_CONTEXT_TERMINAL_PROCESS_SUPPORTED, KEYBINDING_CONTEXT_TERMINAL_IS_OPEN),
-				menu: {
-					id: MenuId.TerminalTabsWidgetContext,
-					group: ContextMenuGroup.Kill
-				},
 				keybinding: {
 					primary: KeyCode.Delete,
 					mac: {
@@ -1619,13 +1506,7 @@ export function registerTerminalActions() {
 			focusNext(accessor);
 		}
 	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalTabInlineActions, {
-		command: {
-			id: TERMINAL_COMMAND_ID.KILL_INSTANCE,
-			title: localize('workbench.action.terminal.killInstance', "Kill Terminal")
-		},
-		group: ContextMenuGroup.Kill
-	});
+
 	registerAction2(class extends Action2 {
 		constructor() {
 			super({
@@ -1641,11 +1522,7 @@ export function registerTerminalActions() {
 					// highest priority when chords are registered afterwards
 					weight: KeybindingWeight.WorkbenchContrib + 1,
 					when: KEYBINDING_CONTEXT_TERMINAL_FOCUS
-				}],
-				menu: {
-					id: MenuId.TerminalContainerContext,
-					group: ContextMenuGroup.Clear
-				}
+				}]
 			});
 		}
 		run(accessor: ServicesAccessor) {
@@ -1669,13 +1546,7 @@ export function registerTerminalActions() {
 			await accessor.get(ITerminalService).showProfileQuickPick('setDefault');
 		}
 	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalToolbarContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.SELECT_DEFAULT_PROFILE,
-			title: { value: localize('workbench.action.terminal.selectDefaultProfile', "Select Default Profile"), original: 'Select Default Profile' }
-		},
-		group: ContextMenuTabsGroup.Configure
-	});
+
 	registerAction2(class extends Action2 {
 		constructor() {
 			super({
@@ -1712,13 +1583,7 @@ export function registerTerminalActions() {
 			await accessor.get(IPreferencesService).openSettings(false, '@feature:terminal');
 		}
 	});
-	MenuRegistry.appendMenuItem(MenuId.TerminalToolbarContext, {
-		command: {
-			id: TERMINAL_COMMAND_ID.CONFIGURE_TERMINAL_SETTINGS,
-			title: localize('workbench.action.terminal.openSettings', "Configure Terminal Settings")
-		},
-		group: ContextMenuTabsGroup.Configure
-	});
+
 	// Some commands depend on platform features
 	if (BrowserFeatures.clipboard.writeText) {
 		registerAction2(class extends Action2 {
@@ -1743,14 +1608,6 @@ export function registerTerminalActions() {
 				await accessor.get(ITerminalService).getActiveInstance()?.copySelection();
 			}
 		});
-		MenuRegistry.appendMenuItem(MenuId.TerminalContainerContext, {
-			command: {
-				id: TERMINAL_COMMAND_ID.COPY_SELECTION,
-				title: localize('workbench.action.terminal.copySelection.short', "Copy")
-			},
-			group: ContextMenuGroup.Edit,
-			order: 1
-		});
 	}
 
 	if (BrowserFeatures.clipboard.readText) {
@@ -1774,14 +1631,6 @@ export function registerTerminalActions() {
 			async run(accessor: ServicesAccessor) {
 				await accessor.get(ITerminalService).getActiveInstance()?.paste();
 			}
-		});
-		MenuRegistry.appendMenuItem(MenuId.TerminalContainerContext, {
-			command: {
-				id: TERMINAL_COMMAND_ID.PASTE,
-				title: localize('workbench.action.terminal.paste.short', "Paste")
-			},
-			group: ContextMenuGroup.Edit,
-			order: 2
 		});
 	}
